+++ Source/SelfService/Web/microservice/microserviceView/healthStatus/dataTableColumns.tsx
@@ -1,67 +1,10 @@
 // Copyright (c) Dolittle. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-import React, { useState, useEffect } from 'react';
+import { GridColDef, GridValueGetterParams } from '@mui/x-data-grid-pro';
 
-import { useSnackbar } from 'notistack';
-
-import { CSVLink } from 'react-csv';
-
-import { getPodLogs } from '../../../api/api';
-
-import { GridColDef, GridValueGetterParams, GridRenderCellParams } from '@mui/x-data-grid-pro';
-import { Button, IconButton, Link } from '@mui/material';
-import { Close, DownloadRounded } from '@mui/icons-material';
-
-import { formatTime, formatStartingDate } from './helpers';
+import { formatTime, formatStartingDate, DownloadLogs } from './helpers';
 import { statusCell } from '../../microserviceStatus';
-
-const DownloadLogs = (params: GridRenderCellParams) => {
-    const [data, setData] = useState({ logs: '' });
-
-    const { enqueueSnackbar, closeSnackbar } = useSnackbar();
-
-    useEffect(() => {
-        getPodLogs(params.row.application, params.row.podName, params.row.containerName).then(data => {
-            setData(data);
-            return;
-        });
-    }, []);
-
-    const logsBlob = new Blob([data.logs], { type: 'text/plain' })
-    const containerImage = params.row.image.replace(':', '/')
-
-    const handleNotification = () => {
-        enqueueSnackbar(`'${containerImage}' logs have been downloaded.`, {
-            variant: 'default',
-            persist: true,
-            anchorOrigin: { horizontal: 'left', vertical: 'bottom' },
-            action: (key) => (
-                <>
-                    <Button
-                        size='small'
-                        onClick={() => alert(`Clicked on action of snackbar with id: ${key}`)}
-                    >
-                        Open
-                    </Button>
-                    <IconButton onClick={() => closeSnackbar(key)}>
-                        <Close />
-                    </IconButton>
-                </>
-            )
-        });
-    };
-
-    return (
-        <Link
-            href={URL.createObjectURL(logsBlob)}
-            download={`${containerImage}.log`}
-            onClick={handleNotification}
-        >
-            <DownloadRounded fontSize='small' sx={{ color: 'text.primary' }} />
-        </Link>
-    )
-};
 
 export const columns: GridColDef[] = [
     {
